Hack for OpRa integration (solving GML issue in submodule)
This small hack is designed to avoid conflict between SIRI an,d NeTEx in their use of GML: they both import the exact same GML extract, with a result of a validation fail if you import both SIRI and NeTEx which is the case for OpRa (OpRA use NeTEx and SIRI as submodules).
This hack changes SIRI so it import the NeTEx GML extract from the NeTEx Git submodule used in OpRa. As the result it DOES NOT VALIDATE as is (needs to be used in OpRa).
This branch is therefore temporary and will be removed as soon as a cleaner solution is provided.
Of cource this branch is absolutely not expected to be used for a Pull Request (it will only be refered by OpRa SIRI submodule)